+++ src/features/dashboard/components/columns/columns.tsx
@@ -1,44 +1,23 @@
 import { ColumnDef } from '@tanstack/react-table';
 import { TableColumnHeader } from '@/components/TableColumnHeader';
-
-export type StockTableData = {
-  holding: {
-    holdingIcon: string;
-    holdingSymbol: string;
-    holdingName: string;
-  };
-  stocksPurchased: number;
-  purchaseValue: {
-    total: number;
-    perShare: number;
-  };
-  currentValue: {
-    total: number;
-    perShare: number;
-  };
-  profit: {
-    absolute: number;
-    percentage: number;
-  };
-  portfolioShare: number;
-};
+import { StockTableData } from './types';
 
 export const columns: ColumnDef<StockTableData>[] = [
   {
     accessorKey: 'holding',
-    header: ({ column }) => {
-      return (
-        <TableColumnHeader
-          column={column}
-          title="Holding"
-          buttonClassName="pl-0"
-          wrapperClassName="text-left"
-        />
-      );
-    },
+    header: ({ column }) => (
+      <TableColumnHeader
+        toggleColumnSorting={() =>
+          column.toggleSorting(column.getIsSorted() === 'asc')
+        }
+        buttonClassName="pl-0"
+        className="text-left"
+      >
+        Holding
+      </TableColumnHeader>
+    ),
     cell: ({ row }) => {
       const holding = row.getValue<StockTableData['holding']>('holding');
-
       return (
         <div className="flex items-center space-x-3 text-left">
           <img
@@ -54,35 +33,35 @@ export const columns: ColumnDef<StockTableData>[] = [
       );
     },
   },
-
   {
     accessorKey: 'stocksPurchased',
-    header: ({ column }) => {
-      return (
-        <TableColumnHeader
-          column={column}
-          title="Počet ks"
-          wrapperClassName="text-center"
-        />
-      );
-    },
-    cell: ({ row }) => {
-      return (
-        <div className="text-center font-medium">
-          {row.getValue('stocksPurchased')}
-        </div>
-      );
-    },
+    header: ({ column }) => (
+      <TableColumnHeader
+        toggleColumnSorting={() =>
+          column.toggleSorting(column.getIsSorted() === 'asc')
+        }
+        className="text-center"
+      >
+        Počet ks
+      </TableColumnHeader>
+    ),
+    cell: ({ row }) => (
+      <div className="text-center font-medium">
+        {row.getValue('stocksPurchased')}
+      </div>
+    ),
   },
-
   {
     accessorKey: 'purchaseValue',
     header: ({ column }) => (
       <TableColumnHeader
-        column={column}
-        title="Kupní hodnota"
-        wrapperClassName="text-center"
-      />
+        toggleColumnSorting={() =>
+          column.toggleSorting(column.getIsSorted() === 'asc')
+        }
+        className="text-center"
+      >
+        Kupní hodnota
+      </TableColumnHeader>
     ),
     cell: ({ row }) => {
       const purchaseValue =
@@ -95,15 +74,17 @@ export const columns: ColumnDef<StockTableData>[] = [
       );
     },
   },
-
   {
     accessorKey: 'currentValue',
     header: ({ column }) => (
       <TableColumnHeader
-        column={column}
-        title="Současná hodnota"
-        wrapperClassName="text-center"
-      />
+        toggleColumnSorting={() =>
+          column.toggleSorting(column.getIsSorted() === 'asc')
+        }
+        className="text-center"
+      >
+        Současná hodnota
+      </TableColumnHeader>
     ),
     cell: ({ row }) => {
       const currentValue =
@@ -116,15 +97,17 @@ export const columns: ColumnDef<StockTableData>[] = [
       );
     },
   },
-
   {
     accessorKey: 'profit',
     header: ({ column }) => (
       <TableColumnHeader
-        column={column}
-        title="Zisk"
-        wrapperClassName="text-center"
-      />
+        toggleColumnSorting={() =>
+          column.toggleSorting(column.getIsSorted() === 'asc')
+        }
+        className="text-center"
+      >
+        Zisk
+      </TableColumnHeader>
     ),
     cell: ({ row }) => {
       const profit = row.getValue<StockTableData['profit']>('profit');
@@ -138,22 +121,21 @@ export const columns: ColumnDef<StockTableData>[] = [
       );
     },
   },
-
   {
     accessorKey: 'portfolioShare',
-    header: ({ column }) => {
-      return (
-        <TableColumnHeader
-          column={column}
-          title="Podíl v portfoliu"
-          wrapperClassName="text-center"
-        />
-      );
-    },
+    header: ({ column }) => (
+      <TableColumnHeader
+        toggleColumnSorting={() =>
+          column.toggleSorting(column.getIsSorted() === 'asc')
+        }
+        className="text-center"
+      >
+        Podíl v portfoliu
+      </TableColumnHeader>
+    ),
     cell: ({ row }) => {
       const portfolioShare = parseFloat(row.getValue('portfolioShare'));
       const formatted = `${portfolioShare.toFixed(2)} %`;
-
       return <div className="text-center font-medium">{formatted}</div>;
     },
   },
