+++ src/features/dashboard/components/columns/types.ts
@@ -1,0 +1,21 @@
+export type StockTableData = {
+  holding: {
+    holdingIcon: string;
+    holdingSymbol: string;
+    holdingName: string;
+  };
+  stocksPurchased: number;
+  purchaseValue: {
+    total: number;
+    perShare: number;
+  };
+  currentValue: {
+    total: number;
+    perShare: number;
+  };
+  profit: {
+    absolute: number;
+    percentage: number;
+  };
+  portfolioShare: number;
+};
