+++ src/features/dashboard/mockData/stockTableData.ts
@@ -1,0 +1,172 @@
+import { StockTableData } from '../components/columns/types';
+
+export const stockTableData: StockTableData[] = [
+  {
+    holding: {
+      holdingIcon: 'https://logo.clearbit.com/apple.com',
+      holdingSymbol: 'AAPL',
+      holdingName: 'Apple Inc.',
+    },
+    stocksPurchased: 15,
+    purchaseValue: {
+      total: 1500,
+      perShare: 100,
+    },
+    currentValue: {
+      total: 1950,
+      perShare: 130,
+    },
+    profit: {
+      absolute: 450,
+      percentage: 30,
+    },
+    portfolioShare: 10,
+  },
+  {
+    holding: {
+      holdingIcon: 'https://logo.clearbit.com/microsoft.com',
+      holdingSymbol: 'MSFT',
+      holdingName: 'Microsoft Corporation',
+    },
+    stocksPurchased: 10,
+    purchaseValue: {
+      total: 2800,
+      perShare: 280,
+    },
+    currentValue: {
+      total: 3100,
+      perShare: 310,
+    },
+    profit: {
+      absolute: 300,
+      percentage: 10.71,
+    },
+    portfolioShare: 12,
+  },
+  {
+    holding: {
+      holdingIcon: 'https://logo.clearbit.com/google.com',
+      holdingSymbol: 'GOOGL',
+      holdingName: 'Alphabet Inc.',
+    },
+    stocksPurchased: 8,
+    purchaseValue: {
+      total: 2000,
+      perShare: 250,
+    },
+    currentValue: {
+      total: 2200,
+      perShare: 275,
+    },
+    profit: {
+      absolute: 200,
+      percentage: 10,
+    },
+    portfolioShare: 15,
+  },
+  {
+    holding: {
+      holdingIcon: 'https://logo.clearbit.com/amazon.com',
+      holdingSymbol: 'AMZN',
+      holdingName: 'Amazon.com, Inc.',
+    },
+    stocksPurchased: 5,
+    purchaseValue: {
+      total: 1650,
+      perShare: 330,
+    },
+    currentValue: {
+      total: 1500,
+      perShare: 300,
+    },
+    profit: {
+      absolute: -150,
+      percentage: -9.09,
+    },
+    portfolioShare: 8,
+  },
+  {
+    holding: {
+      holdingIcon: 'https://logo.clearbit.com/tesla.com',
+      holdingSymbol: 'TSLA',
+      holdingName: 'Tesla, Inc.',
+    },
+    stocksPurchased: 20,
+    purchaseValue: {
+      total: 1200,
+      perShare: 60,
+    },
+    currentValue: {
+      total: 1400,
+      perShare: 70,
+    },
+    profit: {
+      absolute: 200,
+      percentage: 16.67,
+    },
+    portfolioShare: 20,
+  },
+  {
+    holding: {
+      holdingIcon: 'https://logo.clearbit.com/nvidia.com',
+      holdingSymbol: 'NVDA',
+      holdingName: 'NVIDIA Corporation',
+    },
+    stocksPurchased: 12,
+    purchaseValue: {
+      total: 600,
+      perShare: 50,
+    },
+    currentValue: {
+      total: 720,
+      perShare: 60,
+    },
+    profit: {
+      absolute: 120,
+      percentage: 20,
+    },
+    portfolioShare: 10,
+  },
+  {
+    holding: {
+      holdingIcon: 'https://logo.clearbit.com/coca-cola.com',
+      holdingSymbol: 'KO',
+      holdingName: 'The Coca-Cola Company',
+    },
+    stocksPurchased: 30,
+    purchaseValue: {
+      total: 450,
+      perShare: 15,
+    },
+    currentValue: {
+      total: 480,
+      perShare: 16,
+    },
+    profit: {
+      absolute: 30,
+      percentage: 6.67,
+    },
+    portfolioShare: 5,
+  },
+  {
+    holding: {
+      holdingIcon: 'https://logo.clearbit.com/jnj.com',
+      holdingSymbol: 'JNJ',
+      holdingName: 'Johnson & Johnson',
+    },
+    stocksPurchased: 20,
+    purchaseValue: {
+      total: 1800,
+      perShare: 90,
+    },
+    currentValue: {
+      total: 1900,
+      perShare: 95,
+    },
+    profit: {
+      absolute: 100,
+      percentage: 5.56,
+    },
+    portfolioShare: 12,
+  },
+];
