+++ src/features/transactionTable/components/columns/types.ts
@@ -1,0 +1,19 @@
+export type TransactionTableData = {
+  transactionType: string;
+  holding: {
+    holdingIcon: string;
+    holdingSymbol: string;
+    holdingName: string;
+  };
+  transactionDate: string;
+  numberOfStocks: number;
+  transactionValue: {
+    total: number;
+    perShare: number;
+    currency: string;
+  };
+  transactionFee?: {
+    total: number;
+    currency: string;
+  };
+};
